requirements.txt
moviepy>=1.0.3
pims>=0.6.1
PyYAML>=6.0
Pillow>=9.5.0
pytest>=7.4.0
numpy>= 1.25.0rc1
opencv-python>=4.7.0.72